fix(davinci): skip raw text bodies in sfc selector
